requirements.txt
discord.py
python-dotenv
llama_index
llama-index-llms-ollama